refactor: restructured to allow for easier maintenance
