+++ internal/provider/provider.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/integralist/terraform-provider-fastly-framework/internal/provider/resources"
 )
 
 // Ensure FastlyProvider satisfies various provider interfaces.
@@ -61,7 +62,7 @@ func (p *FastlyProvider) Configure(ctx context.Context, req provider.ConfigureRe
 
 func (p *FastlyProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		NewServiceVCLResource,
+		resources.NewServiceVCLResource,
 	}
 }
