+++ internal/provider/resources/docs/service_vcl.md
@@ -1,0 +1,3 @@
+Provides a Fastly Service, representing the configuration for a website, app, API, or anything else to be served through Fastly. A Service encompasses Domains and Backends.
+
+The Service resource requires a domain name configured to direct traffic to the Fastly service. See Fastly's guide on [Adding CNAME Records](https://docs.fastly.com/en/guides/adding-cname-records) on their documentation site for guidance.
